+++ OpenSim/Common/SmoothSegmentedFunction.h
@@ -24,8 +24,14 @@
  * -------------------------------------------------------------------------- */
 #include "osimCommonDLL.h"
 #include "SegmentedQuinticBezierToolkit.h"
+#include <memory>
 
 namespace OpenSim { 
+
+    /**
+    Struct containing the data used by SmoothSegmentedFuncion.
+    */
+    struct SmoothSegmentedFunctionData;
 
     /**
     This class contains the quintic Bezier curves, x(u) and y(u), that have been
@@ -70,6 +76,15 @@ namespace OpenSim {
         ///NaN's
         SmoothSegmentedFunction();
 
+        SmoothSegmentedFunction(const SmoothSegmentedFunction&);
+
+        SmoothSegmentedFunction& operator=(const SmoothSegmentedFunction&);
+
+        ~SmoothSegmentedFunction() noexcept;
+
+        SmoothSegmentedFunction(SmoothSegmentedFunction&&) noexcept;
+
+        SmoothSegmentedFunction& operator=(SmoothSegmentedFunction&&) noexcept;
 
 
        /**Calculates the value of the curve this object represents.
@@ -307,41 +322,10 @@ namespace OpenSim {
        ///@endcond
 
     private:
-       
-        /**Array of spline fit functions X(u) for each Bezier elbow*/
-        SimTK::Array_<SimTK::Spline> _arraySplineUX;        
-        /**Spline fit of the integral of the curve y(x)*/
-        SimTK::Spline _splineYintX;
-        
-        /**Bezier X1,...,Xn control point locations. Control points are 
-        stored in 6x1 vectors in the order above*/
-        SimTK::Array_<SimTK::Vector> _mXVec; 
-        /**Bezier Y1,...,Yn control point locations. Control points are 
-        stored in 6x1 vectors in the order above*/
-        SimTK::Array_<SimTK::Vector> _mYVec; 
 
-        /**The number of quintic Bezier curves that describe the relation*/
-        int _numBezierSections;
+        /**Data required for performing the calculations. **/
+        std::shared_ptr<const SmoothSegmentedFunctionData> _smoothData = nullptr;
 
-        /**The minimum value of the domain*/
-        double _x0;
-        /**The maximum value of the domain*/
-        double _x1;
-        /**The minimum value of the range*/
-        double _y0;
-        /**The maximum value of the range*/
-        double _y1;
-        /**The slope at _x0*/
-        double _dydx0;
-        /**The slope at _x1*/
-        double _dydx1;
-        /**This is the users */
-        bool _computeIntegral;
-
-        /**This variable, when true, indicates that the user wants the integral
-        from left to right (x0 to x1). If it is false, the integral from right
-        to left (x1 to x0) is computed*/
-        bool _intx0x1;
         /**The name of the function**/
         std::string _name;